move method to utils
so I can add to eventually add to gui
